refactor(ratelimiter.spec): shift to own test file
